Adding WAV Generation
Setting up some scaffolding for creating a raw WAV stream to pass to the audio out device on transmit, instead of relying on the unsteady Thread.Sleep method that won't hold up when a machine has trouble keeping up. Also, no, I didn't create this WAVE library all in one day -- it's from a separate project of mine and I'm migrating it over to this one. with a lot of cleanup Perhaps I could package it if I make it modular enough.