I'll Be Damned
Problem was that I was using a "new WaveAudioFormat" for the WaveStream constructor. Fixed that now to pull the first one from a stream instead of parameterizing it. Regardless, I removed all Linq stuff anyway; not a big deal. Testing went well for 16 and 24 bitrates, but DESTROYED my speakers on 8-bit. 32-bit won't even load the application -- might look into it...