WPM Working Properly
Now that WPM works just fine, it's back to the task of syncing the audio with the text. Debug shows the text I'd expect, but that's not what's happening. Going to break for now and checkpoint myself. I think this has to do with THREAD.SLEEP being awful for timing at small intervals, but it doesn't even work at 5 WPM...