TX WORKING!
It's at the right speed, and the cursor/live view syncs up pretty well! I can't believe I missed that sneaky leftover index++ at the bottom, which was causing a double-incrementation of the current for-loop index. It was an artifact from the NAudio player where I was only using INDEX for the progress bar...