Clean-up & Resurrection
Massive refactor of the project. Moved to .NET Core 5.0 for the future of development. Removed the static files that were acting as the ToneBaker audio library. This project now wholly depends on the ToneBaker.dll file instead of halfheartedly using CS file copies. Of course developing this locally means some fixups are in order and eventually I'll get around to a fix, but wanted to establish the refactored BASELINE for the project first.